Create test utils module
