+++ settings.gradle
@@ -3,3 +3,4 @@ include 'tempest'
 include ':samples:guides'
 include ':samples:musiclibrary'
 include ':samples:urlshortener'
+include 'test-utils'
