+++ test-utils/build.gradle
@@ -3,21 +3,12 @@ apply plugin: 'kotlin'
 apply plugin: 'com.vanniktech.maven.publish'
 
 dependencies {
-  api dep.awsDynamodb
   api dep.findbugsJsr305
-  implementation project(":tempest-internal")
   implementation dep.kotlinReflection
   implementation dep.kotlinStdLib
-  implementation dep.okio
 
-  testImplementation project(":samples:musiclibrary")
-  testImplementation project(":samples:urlshortener")
-  testImplementation project(":test-utils")
-  testImplementation dep.miskAwsDynamodbTesting
   testImplementation dep.assertj
   testImplementation dep.miskTesting
   testImplementation dep.junitApi
   testImplementation dep.junitEngine
 }
-
-apply from: "$rootDir/gradle-mvn-publish.gradle"
