+++ test-utils/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=test-utils
+POM_NAME=test-utils
+POM_DESCRIPTION=test-utils
+POM_PACKAGING=jar
