+++ test-utils/src/main/kotlin/app/cash/tempest/musiclibrary/TestEntities.kt
@@ -191,3 +191,16 @@ val LOCKDOWN_SINGLE = Album(
     Track("Lockdown'", Duration.parse("PT3M18S"))
   )
 )
+
+data class Album(
+  val album_token: String,
+  val album_title: String,
+  val artist_name: String,
+  val release_date: LocalDate,
+  val genre_name: String,
+  val tracks: List<Track>
+) {
+  val trackTitles = tracks.map { it.track_title }
+}
+
+data class Track(val track_title: String, val run_length: Duration)
